adding useToolbar composable
